Update
This is capable of building a 1000,2000 floor for some house, and it can build a table, of your scalar choice somewhere comfortably within the floor's bounds. Next I will work on walls, but that will be hard with out the camer seeing as I wont be able to get inside the walls.

-I developed a simulation of myself and that is who wrote this code and message